requirements.txt
flask
pymupdf
python-docx
gunicorn
